requirements.txt
amqp==1.4.9
anyjson==0.3.3
awscli==1.7.41
awsebcli==3.7.7
bcdoc==0.16.0
billiard==3.3.0.23
blessed==1.9.5
boto==2.41.0
botocore==1.4.37
cement==2.8.2
colorama==0.3.7
defusedxml==0.4.1
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.3
django-allauth==0.30.0
django-appconf==1.0.1
django-argonauts==1.1.3
django-braces==1.10.0
django-compressor==1.5
django-cors-headers==1.1.0
django-cors-middleware==1.3.1
django-debug-toolbar==1.4
django-filter==1.0.0
django-grappelli==2.9.1
django-multiselectfield==0.1.6.dev1
django-model-utils==2.5
django-oauth-toolkit==0.10.0
django-rest-auth==0.9.0
django-rest-framework-social-oauth2==1.0.4
djangorestframework==3.5.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
drf-nested-routers==0.9.0
ecdsa==0.13
gunicorn==19.6.0
jmespath==0.9.0
kombu==3.0.35
langdetect==1.0.5
oauthlib==1.0.3
paramiko==1.17.1
pathspec==0.3.4
prettytable==0.7.2
psycopg2==2.6.1
pyasn1==0.1.9
pycrypto==2.6.1
PyJWT==1.4.2
python-dateutil==2.5.3
python-social-auth==0.2.21
python3-openid==3.0.10
PyYAML==3.11
rcssmin==1.0.6
requests==2.12.1
requests-oauthlib==0.7.0
rjsmin==1.0.12
rsa==3.1.4
semantic-version==2.5.0
six==1.10.0
South==1.0.2
sqlparse==0.1.19
ssh==1.8.0
static3==0.6.1
texttable==0.8.4
wcwidth==0.1.7
websocket-client==0.37.0
whitenoise==3.2.1
